Share project file fragment cleanup
